Refactor campaign and user action payloads, update import paths, and remove unused files
- Updated import paths for user alias generation and user payloads.
- Refactored campaign payloads to improve clarity and consistency.
- Removed obsolete files related to in-app campaigns and user actions.
- Adjusted test files to reflect changes in payload structure and import paths.
