+++ src/api/models/mobile.users.model.ts
@@ -23,13 +23,6 @@ export interface User {
   phone: string;
 }
 
-export interface UserAction {
-  guid: string;
-  key: string;
-  occurred_at_array: string[];
-  type: string;
-}
-
 export interface UpdateMobileUserPayload {
   alias: string;
   current_location: number[];
@@ -38,4 +31,18 @@ export interface UpdateMobileUserPayload {
   screen_records: never[];
   user: User;
   user_actions: UserAction[];
+}
+export interface UserAction {
+  guid: string;
+  key: InAppEvents;
+  occurred_at_array: string[];
+  type: string;
+}
+
+export enum InAppEvents {
+  IN_APP_DELIVERY = 'in_app_delivery',
+  IN_APP_IMPRESSION = 'in_app_impression',
+  IN_APP_DISMISS = 'in_app_dismiss',
+  IN_APP_BUTTON_CLICK_ONE = 'in_app_button_click_one',
+  IN_APP_BUTTON_CLICK_TWO = 'in_app_button_click_two'
 }
