+++ src/api/test-data/mobile/sessions/start-session.payload.ts
@@ -1,15 +1,15 @@
 import { StartMobileSessionPayload } from '@_src/api/models/mobile.sessions.model';
 import { faker } from '@faker-js/faker/locale/en';
 
-export const startMobileSessionPayload: StartMobileSessionPayload = {
+export const startMobileSessionInAppPayload: StartMobileSessionPayload = {
   alias: '',
   device: {
     type: 'android',
     token: faker.string.uuid(),
     bluetooth_state: 'on',
     os_version: 'ios',
-    sdk_version: '2.13.8',
-    app_version: '2.13.8-prod',
+    sdk_version: '4.7.0',
+    app_version: '4.7.0.9',
     language: 'en',
     location_permission: true,
     push_permission: true,
