--- src/api/test-data/mobile/users/carrd/feed-post-impression-payload.ts
@@ -1,8 +1,0 @@
-import { UserAction } from '@_src/api/models/mobile.users.model';
-
-export const feedPostFrontImpressionAction: UserAction = {
-  guid: '',
-  key: 'card_front_impression',
-  occurred_at_array: ['1733154276'],
-  type: 'campaign'
-};
