--- src/api/test-data/mobile/users/inapp/in-app-delivery.payload.ts
@@ -1,8 +1,0 @@
-import { UserAction } from '@_src/api/models/mobile.users.model';
-
-export const inAppDeliveryAction: UserAction = {
-  guid: '',
-  key: 'in_app_delivery',
-  occurred_at_array: ['1730990106'],
-  type: 'campaign'
-};
