--- src/api/test-data/mobile/users/inapp/in-app-dismiss.payload.ts
@@ -1,9 +1,0 @@
-// src/api/test-data/user-actions/in-app-dismiss.ts
-import { UserAction } from '@_src/api/models/mobile.users.model';
-
-export const inAppDismissAction: UserAction = {
-  guid: '',
-  key: 'in_app_dismiss',
-  occurred_at_array: ['1730990108'],
-  type: 'campaign'
-};
