--- src/api/test-data/mobile/users/inapp/in-app-impression.payload.ts
@@ -1,8 +1,0 @@
-import { UserAction } from '@_src/api/models/mobile.users.model';
-
-export const inAppImpressionAction: UserAction = {
-  guid: '',
-  key: 'in_app_impression',
-  occurred_at_array: ['1730990107'],
-  type: 'campaign'
-};
