+++ tests/api/admin.register.api.spec.ts
