--- tests/api/campaign.inapp.api.spec.ts
@@ -1,238 +1,0 @@
-import {
-  API_E2E_ACCESS_TOKEN_ADMIN,
-  API_E2E_APP_ID,
-  SUPER_ADMIN_ACCESS_TOKEN
-} from '@_config/env.config';
-import { getSdkCredentials } from '@_src/api/factories/app.api.factory';
-import {
-  createCampaignWithApi,
-  deleteCampaignWithApi,
-  getCampaignsWithApi
-} from '@_src/api/factories/campaigns.api.factory';
-import { startMobileSessionsWithApi } from '@_src/api/factories/mobile.sessions.api.factory';
-import { updateMobileUserWithApi } from '@_src/api/factories/mobile.users.api.factory';
-import { createSegmentWithApi } from '@_src/api/factories/segments.api.factory';
-import { getCampaignStatsWithApi } from '@_src/api/factories/stats.api.factory';
-import { getAllUsersWithApi } from '@_src/api/factories/users.api.factory';
-import {
-  APIE2ELoginUserModel,
-  APIE2ETokenSDKModel
-} from '@_src/api/models/admin.model';
-import { createCampaignPayloadInAppLarge } from '@_src/api/test-data/cms/campaign/create-inapp-large-campaign.payload';
-import { createSegmentAllUsersPayload } from '@_src/api/test-data/cms/segment/create-segment-all-users.payload';
-import { startMobileSessionPayload } from '@_src/api/test-data/mobile/sessions/mobile.sessions.payload';
-import { inAppDeliveryAction } from '@_src/api/test-data/mobile/users/inapp/in-app-delivery.payload';
-import { inAppDismissAction } from '@_src/api/test-data/mobile/users/inapp/in-app-dismiss.payload';
-import { inAppImpressionAction } from '@_src/api/test-data/mobile/users/inapp/in-app-impression.payload';
-import { updateMobileUserPayload } from '@_src/api/test-data/mobile/users/update/mobile.users.update.payload';
-import {
-  deleteAllCampaigns,
-  deleteAllSegments,
-  deleteAllUsers,
-  importRandomUsers
-} from '@_src/api/utils/data.manager.util';
-import { expect, test } from '@_src/ui/fixtures/merge.fixture';
-
-test.describe('In-App Campaign Tests', () => {
-  let APIE2ETokenSDKModel: APIE2ETokenSDKModel;
-
-  const APIE2ELoginUserModel: APIE2ELoginUserModel = {
-    apiE2EAccessTokenAdmin: `${API_E2E_ACCESS_TOKEN_ADMIN}`,
-    apiE2EAccessTokenSuperAdmin: `${SUPER_ADMIN_ACCESS_TOKEN}`,
-    apiE2EAppId: `${API_E2E_APP_ID}`
-  };
-
-  test.beforeAll(async ({ request }) => {
-    const sdkCredentialsResponse = await getSdkCredentials(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      APIE2ELoginUserModel.apiE2EAppId
-    );
-    const sdkCredentials = await sdkCredentialsResponse.json();
-
-    APIE2ETokenSDKModel = {
-      apiE2EAccessTokenSdk: sdkCredentials.access_token
-    };
-  });
-
-  test.beforeEach(async ({ request }) => {
-    await deleteAllUsers(request, APIE2ELoginUserModel.apiE2EAccessTokenAdmin);
-
-    await deleteAllSegments(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
-    );
-
-    await deleteAllCampaigns(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
-    );
-  });
-
-  test('should create and delete an In-App Large campaign', async ({
-    request
-  }) => {
-    const numberOfUsers = 5;
-
-    await importRandomUsers(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      APIE2ELoginUserModel.apiE2EAppId,
-      numberOfUsers
-    );
-
-    // Create Segment
-    const createSegmentResponse = await createSegmentWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      createSegmentAllUsersPayload
-    );
-    const createSegmentResponseJson = await createSegmentResponse.json();
-
-    // Create Campaign
-    createCampaignPayloadInAppLarge.segment_ids = [
-      createSegmentResponseJson.segment.id
-    ];
-    const createCampaignResponse = await createCampaignWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      createCampaignPayloadInAppLarge
-    );
-    const createCampaignResponseJson = await createCampaignResponse.json();
-
-    // Assert Campaign Created
-    expect(createCampaignResponseJson.name).toBe(
-      createCampaignPayloadInAppLarge.name
-    );
-
-    // Get Campaigns
-    const getCampaignsResponse = await getCampaignsWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
-    );
-    const getCampaignsResponseJson = await getCampaignsResponse.json();
-
-    const createdCampaign = getCampaignsResponseJson.data.find(
-      (campaign: { id: string }) =>
-        campaign.id === createCampaignResponseJson.id
-    );
-
-    // Assert Campaign Exists
-    expect(createdCampaign).toBeDefined();
-    expect(createdCampaign.name).toBe(createCampaignPayloadInAppLarge.name);
-
-    // Delete Campaign
-    await deleteCampaignWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      createdCampaign.id
-    );
-
-    // Verify Deletion
-    const getCampaignsResponseAfterDeletion = await getCampaignsWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
-    );
-    const getCampaignsResponseJsonAfterDeletion =
-      await getCampaignsResponseAfterDeletion.json();
-    expect(getCampaignsResponseJsonAfterDeletion.data.length).toBe(0);
-  });
-
-  test('should create an In-App Large campaign with a dismiss button and update mobile session user to click it', async ({
-    request
-  }) => {
-    const numberOfUsers = 1;
-
-    await importRandomUsers(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      APIE2ELoginUserModel.apiE2EAppId,
-      numberOfUsers
-    );
-
-    // Create Segment
-    const createSegmentResponse = await createSegmentWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      createSegmentAllUsersPayload
-    );
-    const createSegmentResponseJson = await createSegmentResponse.json();
-
-    // Create Campaign
-    createCampaignPayloadInAppLarge.segment_ids = [
-      createSegmentResponseJson.segment.id
-    ];
-    const createCampaignResponse = await createCampaignWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      createCampaignPayloadInAppLarge
-    );
-    const createCampaignResponseJson = await createCampaignResponse.json();
-
-    // Assert Campaign Created
-    expect(createCampaignResponseJson.name).toBe(
-      createCampaignPayloadInAppLarge.name
-    );
-
-    const getUsersResponse = await getAllUsersWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
-    );
-    const getUsersResponseJson = await getUsersResponse.json();
-
-    startMobileSessionPayload.alias = getUsersResponseJson.data[0].alias;
-    updateMobileUserPayload.alias = getUsersResponseJson.data[0].alias;
-
-    // Start Mobile Session
-    await startMobileSessionsWithApi(
-      request,
-      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
-      startMobileSessionPayload
-    );
-
-    // Update Mobile User
-    const userActions = [
-      inAppDeliveryAction,
-      inAppImpressionAction,
-      inAppDismissAction
-    ];
-
-    userActions.forEach((action) => {
-      action.guid = createCampaignResponseJson.guid;
-    });
-
-    await updateMobileUserWithApi(
-      request,
-      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
-      { ...updateMobileUserPayload, user_actions: userActions }
-    );
-
-    const getCampaignStatsWithWaitResponse = await getCampaignStatsWithApi(
-      request,
-      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
-      createCampaignResponseJson.id,
-      1
-    );
-
-    const getCampaignStatsWithWaitResponseJson =
-      await getCampaignStatsWithWaitResponse.json();
-
-    // Assert Validate Response Body
-    expect(getCampaignStatsWithWaitResponseJson).toHaveProperty('in_app');
-    expect(getCampaignStatsWithWaitResponseJson.in_app.send).toHaveProperty(
-      'total_uniq',
-      1
-    );
-    expect(getCampaignStatsWithWaitResponseJson.in_app.delivery).toHaveProperty(
-      'total_uniq',
-      1
-    );
-    expect(getCampaignStatsWithWaitResponseJson.in_app.dismiss).toHaveProperty(
-      'total_uniq',
-      1
-    );
-    expect(
-      getCampaignStatsWithWaitResponseJson.in_app.impression
-    ).toHaveProperty('total_uniq', 1);
-  });
-});
