+++ tests/api/campaign.large.inapp.api.spec.ts
@@ -1,0 +1,581 @@
+import {
+  API_E2E_ACCESS_TOKEN_ADMIN,
+  API_E2E_APP_ID,
+  SUPER_ADMIN_ACCESS_TOKEN
+} from '@_config/env.config';
+import { getSdkCredentials } from '@_src/api/factories/app.api.factory';
+import { createCampaignWithApi } from '@_src/api/factories/campaigns.api.factory';
+import { createDeeplinkWithApi } from '@_src/api/factories/deeplinks.api.factory';
+import { updateDeeplinkWithApi } from '@_src/api/factories/deeplinks.api.factory';
+import { startMobileSessionsWithApi } from '@_src/api/factories/mobile.sessions.api.factory';
+import { updateMobileUserWithApi } from '@_src/api/factories/mobile.users.api.factory';
+import { createSegmentWithApi } from '@_src/api/factories/segments.api.factory';
+import { getCampaignStatsWithApi } from '@_src/api/factories/stats.api.factory';
+import { getAllUsersWithApi } from '@_src/api/factories/users.api.factory';
+import {
+  APIE2ELoginUserModel,
+  APIE2ETokenSDKModel
+} from '@_src/api/models/admin.model';
+import { InAppEvents } from '@_src/api/models/mobile.users.model';
+import {
+  createCampaignInAppLargeButtonWithDismiss,
+  createCampaignInAppLargeButtonWithUrl
+} from '@_src/api/test-data/cms/campaign/create-inapp-campaign.payload';
+import { createSegmentAllUsersPayload } from '@_src/api/test-data/cms/segment/create-segment-all-users.payload';
+import { startMobileSessionInAppPayload } from '@_src/api/test-data/mobile/sessions/start-session.payload';
+import {
+  updateMobileInAppUserPayload,
+  userActions
+} from '@_src/api/test-data/mobile/update/update-user.payload';
+import { apiUrls } from '@_src/api/utils/api.util';
+import {
+  deleteAllCampaigns,
+  deleteAllDeeplinks,
+  deleteAllSegments,
+  deleteAllUsers,
+  importRandomUsers
+} from '@_src/api/utils/data.manager.util';
+import { expect, test } from '@_src/ui/fixtures/merge.fixture';
+
+test.describe('Large In-App Campaign', () => {
+  let APIE2ETokenSDKModel: APIE2ETokenSDKModel;
+
+  const APIE2ELoginUserModel: APIE2ELoginUserModel = {
+    apiE2EAccessTokenAdmin: `${API_E2E_ACCESS_TOKEN_ADMIN}`,
+    apiE2EAccessTokenSuperAdmin: `${SUPER_ADMIN_ACCESS_TOKEN}`,
+    apiE2EAppId: `${API_E2E_APP_ID}`
+  };
+
+  test.beforeAll(async ({ request }) => {
+    const sdkCredentialsResponse = await getSdkCredentials(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      APIE2ELoginUserModel.apiE2EAppId
+    );
+    const sdkCredentials = await sdkCredentialsResponse.json();
+
+    APIE2ETokenSDKModel = {
+      apiE2EAccessTokenSdk: sdkCredentials.access_token
+    };
+  });
+
+  test.beforeEach(async ({ request }) => {
+    await deleteAllUsers(request, APIE2ELoginUserModel.apiE2EAccessTokenAdmin);
+
+    await deleteAllSegments(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
+    );
+
+    await deleteAllCampaigns(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
+    );
+  });
+  test('should create an In-App Large campaign with a URL button and click it', async ({
+    request
+  }) => {
+    const numberOfUsers = 1;
+
+    await importRandomUsers(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      APIE2ELoginUserModel.apiE2EAppId,
+      numberOfUsers
+    );
+
+    // Create Segment
+    const createSegmentResponse = await createSegmentWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createSegmentAllUsersPayload
+    );
+    const createSegmentResponseJson = await createSegmentResponse.json();
+
+    // Create Campaign
+    createCampaignInAppLargeButtonWithUrl.segment_ids = [
+      createSegmentResponseJson.segment.id
+    ];
+    const createCampaignResponse = await createCampaignWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createCampaignInAppLargeButtonWithUrl
+    );
+    const createCampaignResponseJson = await createCampaignResponse.json();
+
+    // Assert Campaign Created
+    expect(createCampaignResponseJson.name).toBe(
+      createCampaignInAppLargeButtonWithUrl.name
+    );
+
+    const getUsersResponse = await getAllUsersWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
+    );
+    const getUsersResponseJson = await getUsersResponse.json();
+
+    startMobileSessionInAppPayload.alias = getUsersResponseJson.data[0].alias;
+    updateMobileInAppUserPayload.alias = getUsersResponseJson.data[0].alias;
+
+    // Start Mobile Session
+    await startMobileSessionsWithApi(
+      request,
+      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
+      startMobileSessionInAppPayload
+    );
+
+    const updateMobileInAppUserPayloadWithActions = {
+      ...updateMobileInAppUserPayload,
+      user_actions: [
+        {
+          ...userActions[InAppEvents.IN_APP_DELIVERY],
+          guid: createCampaignResponseJson.guid
+        },
+        {
+          ...userActions[InAppEvents.IN_APP_IMPRESSION],
+          guid: createCampaignResponseJson.guid
+        },
+        {
+          ...userActions[InAppEvents.IN_APP_BUTTON_CLICK_ONE],
+          guid: createCampaignResponseJson.guid
+        }
+      ]
+    };
+
+    await updateMobileUserWithApi(
+      request,
+      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
+      updateMobileInAppUserPayloadWithActions
+    );
+
+    const getCampaignStatsWithWaitResponse = await getCampaignStatsWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createCampaignResponseJson.id,
+      1
+    );
+
+    const getCampaignStatsWithWaitResponseJson =
+      await getCampaignStatsWithWaitResponse.json();
+
+    // Assert Validate Response Body
+    expect(getCampaignStatsWithWaitResponseJson).toHaveProperty('in_app');
+    expect(getCampaignStatsWithWaitResponseJson.in_app.send).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.delivery).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.bounce).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.error).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.dismiss).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(
+      getCampaignStatsWithWaitResponseJson.in_app.button_click_one
+    ).toHaveProperty('total_uniq', 1);
+    expect(getCampaignStatsWithWaitResponseJson.in_app.clicks).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(
+      getCampaignStatsWithWaitResponseJson.in_app.impression
+    ).toHaveProperty('total_uniq', 1);
+  });
+
+  test('should create an In-App Large campaign with a Deeplink button and click it', async ({
+    request
+  }) => {
+    const numberOfUsers = 1;
+
+    await importRandomUsers(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      APIE2ELoginUserModel.apiE2EAppId,
+      numberOfUsers
+    );
+
+    // Create Segment
+    const createSegmentResponse = await createSegmentWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createSegmentAllUsersPayload
+    );
+    const createSegmentResponseJson = await createSegmentResponse.json();
+
+    await deleteAllDeeplinks(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
+    );
+
+    const createDeeplinkResponse = await createDeeplinkWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      {
+        nickname: 'Plawright',
+        target: `${apiUrls.campaigns.v2.base}`
+      }
+    );
+
+    const updateDeeplinkResponse = await updateDeeplinkWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createDeeplinkResponse.id,
+      {
+        nickname: 'Plawright Campaign Deeplink',
+        target: `${apiUrls.campaigns.v2.base}`
+      }
+    );
+
+    // Preparing payload for campaign creation
+    createCampaignInAppLargeButtonWithUrl.segment_ids = [
+      createSegmentResponseJson.segment.id
+    ];
+    createCampaignInAppLargeButtonWithUrl.in_app_notification.large.call_to_action.buttons[0].destination =
+      updateDeeplinkResponse.id;
+
+    const createCampaignResponse = await createCampaignWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createCampaignInAppLargeButtonWithUrl
+    );
+    const createCampaignResponseJson = await createCampaignResponse.json();
+
+    // Assert Campaign Created
+    expect(createCampaignResponseJson.name).toBe(
+      createCampaignInAppLargeButtonWithUrl.name
+    );
+
+    const getUsersResponse = await getAllUsersWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
+    );
+    const getUsersResponseJson = await getUsersResponse.json();
+
+    startMobileSessionInAppPayload.alias = getUsersResponseJson.data[0].alias;
+    updateMobileInAppUserPayload.alias = getUsersResponseJson.data[0].alias;
+
+    // Start Mobile Session
+    await startMobileSessionsWithApi(
+      request,
+      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
+      startMobileSessionInAppPayload
+    );
+
+    const updateMobileInAppUserPayloadWithActions = {
+      ...updateMobileInAppUserPayload,
+      user_actions: [
+        {
+          ...userActions[InAppEvents.IN_APP_DELIVERY],
+          guid: createCampaignResponseJson.guid
+        },
+        {
+          ...userActions[InAppEvents.IN_APP_IMPRESSION],
+          guid: createCampaignResponseJson.guid
+        },
+        {
+          ...userActions[InAppEvents.IN_APP_BUTTON_CLICK_ONE],
+          guid: createCampaignResponseJson.guid
+        }
+      ]
+    };
+
+    await updateMobileUserWithApi(
+      request,
+      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
+      updateMobileInAppUserPayloadWithActions
+    );
+
+    const getCampaignStatsWithWaitResponse = await getCampaignStatsWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createCampaignResponseJson.id,
+      1
+    );
+
+    const getCampaignStatsWithWaitResponseJson =
+      await getCampaignStatsWithWaitResponse.json();
+
+    // Assert Validate Response Body
+    expect(getCampaignStatsWithWaitResponseJson).toHaveProperty('in_app');
+    expect(getCampaignStatsWithWaitResponseJson.in_app.send).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.delivery).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.bounce).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.error).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.dismiss).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(
+      getCampaignStatsWithWaitResponseJson.in_app.button_click_one
+    ).toHaveProperty('total_uniq', 1);
+    expect(getCampaignStatsWithWaitResponseJson.in_app.clicks).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(
+      getCampaignStatsWithWaitResponseJson.in_app.impression
+    ).toHaveProperty('total_uniq', 1);
+  });
+
+  test('should create an In-App Large campaign with a dismiss button and click it', async ({
+    request
+  }) => {
+    const numberOfUsers = 1;
+
+    await importRandomUsers(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      APIE2ELoginUserModel.apiE2EAppId,
+      numberOfUsers
+    );
+
+    // Create Segment
+    const createSegmentResponse = await createSegmentWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createSegmentAllUsersPayload
+    );
+    const createSegmentResponseJson = await createSegmentResponse.json();
+
+    // Create Campaign
+    createCampaignInAppLargeButtonWithDismiss.segment_ids = [
+      createSegmentResponseJson.segment.id
+    ];
+    const createCampaignResponse = await createCampaignWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createCampaignInAppLargeButtonWithDismiss
+    );
+    const createCampaignResponseJson = await createCampaignResponse.json();
+
+    // Assert Campaign Created
+    expect(createCampaignResponseJson.name).toBe(
+      createCampaignInAppLargeButtonWithDismiss.name
+    );
+
+    const getUsersResponse = await getAllUsersWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
+    );
+    const getUsersResponseJson = await getUsersResponse.json();
+
+    startMobileSessionInAppPayload.alias = getUsersResponseJson.data[0].alias;
+    updateMobileInAppUserPayload.alias = getUsersResponseJson.data[0].alias;
+
+    // Start Mobile Session
+    await startMobileSessionsWithApi(
+      request,
+      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
+      startMobileSessionInAppPayload
+    );
+
+    const updateMobileInAppUserPayloadWithActions = {
+      ...updateMobileInAppUserPayload,
+      user_actions: [
+        {
+          ...userActions[InAppEvents.IN_APP_DELIVERY],
+          guid: createCampaignResponseJson.guid
+        },
+        {
+          ...userActions[InAppEvents.IN_APP_IMPRESSION],
+          guid: createCampaignResponseJson.guid
+        },
+        {
+          ...userActions[InAppEvents.IN_APP_DISMISS],
+          guid: createCampaignResponseJson.guid
+        }
+      ]
+    };
+
+    await updateMobileUserWithApi(
+      request,
+      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
+      updateMobileInAppUserPayloadWithActions
+    );
+
+    const getCampaignStatsWithWaitResponse = await getCampaignStatsWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createCampaignResponseJson.id,
+      1
+    );
+
+    const getCampaignStatsWithWaitResponseJson =
+      await getCampaignStatsWithWaitResponse.json();
+
+    // Assert Validate Response Body
+    expect(getCampaignStatsWithWaitResponseJson).toHaveProperty('in_app');
+    expect(getCampaignStatsWithWaitResponseJson.in_app.send).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.delivery).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.bounce).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.error).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.dismiss).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(
+      getCampaignStatsWithWaitResponseJson.in_app.button_click_one
+    ).toHaveProperty('total_uniq', 0);
+    expect(getCampaignStatsWithWaitResponseJson.in_app.clicks).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(
+      getCampaignStatsWithWaitResponseJson.in_app.impression
+    ).toHaveProperty('total_uniq', 1);
+  });
+
+  test('should create an In-App Large campaign with 2 users but only 1 user clicks the button', async ({
+    request
+  }) => {
+    const numberOfUsers = 2;
+
+    await importRandomUsers(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      APIE2ELoginUserModel.apiE2EAppId,
+      numberOfUsers
+    );
+
+    // Create Segment
+    const createSegmentResponse = await createSegmentWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createSegmentAllUsersPayload
+    );
+    const createSegmentResponseJson = await createSegmentResponse.json();
+
+    // Create Campaign
+    createCampaignInAppLargeButtonWithUrl.segment_ids = [
+      createSegmentResponseJson.segment.id
+    ];
+    const createCampaignResponse = await createCampaignWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createCampaignInAppLargeButtonWithUrl
+    );
+    const createCampaignResponseJson = await createCampaignResponse.json();
+
+    // Assert Campaign Created
+    expect(createCampaignResponseJson.name).toBe(
+      createCampaignInAppLargeButtonWithUrl.name
+    );
+
+    const getUsersResponse = await getAllUsersWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin
+    );
+    const getUsersResponseJson = await getUsersResponse.json();
+
+    startMobileSessionInAppPayload.alias = getUsersResponseJson.data[0].alias;
+    updateMobileInAppUserPayload.alias = getUsersResponseJson.data[0].alias;
+
+    // Start Mobile Session
+    await startMobileSessionsWithApi(
+      request,
+      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
+      startMobileSessionInAppPayload
+    );
+
+    const updateMobileInAppUserPayloadWithActions = {
+      ...updateMobileInAppUserPayload,
+      user_actions: [
+        {
+          ...userActions[InAppEvents.IN_APP_DELIVERY],
+          guid: createCampaignResponseJson.guid
+        },
+        {
+          ...userActions[InAppEvents.IN_APP_IMPRESSION],
+          guid: createCampaignResponseJson.guid
+        },
+        {
+          ...userActions[InAppEvents.IN_APP_BUTTON_CLICK_ONE],
+          guid: createCampaignResponseJson.guid
+        }
+      ]
+    };
+
+    await updateMobileUserWithApi(
+      request,
+      APIE2ETokenSDKModel.apiE2EAccessTokenSdk,
+      updateMobileInAppUserPayloadWithActions
+    );
+
+    const getCampaignStatsWithWaitResponse = await getCampaignStatsWithApi(
+      request,
+      APIE2ELoginUserModel.apiE2EAccessTokenAdmin,
+      createCampaignResponseJson.id,
+      1
+    );
+
+    const getCampaignStatsWithWaitResponseJson =
+      await getCampaignStatsWithWaitResponse.json();
+
+    // Assert Validate Response Body
+    expect(getCampaignStatsWithWaitResponseJson).toHaveProperty('in_app');
+    expect(getCampaignStatsWithWaitResponseJson.in_app.send).toHaveProperty(
+      'total_uniq',
+      2
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.delivery).toHaveProperty(
+      'total_uniq',
+      2
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.bounce).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.error).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(getCampaignStatsWithWaitResponseJson.in_app.dismiss).toHaveProperty(
+      'total_uniq',
+      0
+    );
+    expect(
+      getCampaignStatsWithWaitResponseJson.in_app.button_click_one
+    ).toHaveProperty('total_uniq', 1);
+    expect(getCampaignStatsWithWaitResponseJson.in_app.clicks).toHaveProperty(
+      'total_uniq',
+      1
+    );
+    expect(
+      getCampaignStatsWithWaitResponseJson.in_app.impression
+    ).toHaveProperty('total_uniq', 1);
+  });
+});
